requirements.txt
# Core
aiohttp>=3.9.0
aiofiles>=23.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
pydantic>=2.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.0.0
stem>=1.8.0
tldextract>=5.0.0
fake-useragent>=1.4.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Tor SOCKS proxy support
aiohttp-socks>=0.8.0

# AI digest
anthropic>=0.28.0

# Engine upgrades — TLS fingerprint impersonation
curl_cffi>=0.7.0

# FastAPI server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.6.0
websockets>=12.0

# Infrastructure X-ray
dnspython>=2.5.0
python-whois>=0.9.0

# Export
openpyxl>=3.1.0

# Alerting & HTTP client
httpx>=0.27.0
